AMP-23740: Moved some old dashboard classes to esrigis. Little refactoring.
