--- amp/TEMPLATE/ampTemplate/css_2/visualization_yui_tabs.css
@@ -1,175 +1,0 @@
-
-#demo .yui-nav {
-}
-
-#demo .yui-nav li {
-    margin-right: 0;
-}
-
-#demo .yui-nav li.selected a {
-    border-bottom:1px solid white;
-    margin-bottom: -1px;
-}
-
-#demo .yui-content {
-	padding: 1em 1em 1em 1em;
-}
-
-#demo .yui-nav li a {
-    float: left;
-    cursor: pointer;
-    text-decoration: none;
-    font-size: 12px;
-    font-weight: bold;
-    color: #376091;
-    margin: 0pt 0px 0pt 0pt;
-    /* background:#3754A1 url(tableftcornerunsel.gif) no-repeat scroll left top;*/
-    background-color: #f7f7f7;
-    border-top: 1px solid #D0D0D0;
-    border-left: 1px solid #D0D0D0;
-    border-right: 1px solid #D0D0D0;
-    margin-right: 7px;
-}
-
-#demo li.selected a {
-    float: left;
-    cursor: pointer;
-    text-decoration: none;
-    font-size: 12px;
-    font-weight: bold;
-    /* background:#222E5D url(tableftcorner.gif) no-repeat scroll left top; */
-    color: #FF6000;
-    background-color: #FFFFFF;
-}
-
-#demo .longTab {
-    width: 100%;
-    float: none;
-    font-size: 12px;
-    color: white;
-    font-weight: bold;
-    height: 25px;
-    padding-top: 5px;
-    background-color: #222E5D
-}
-
-#demo li a div {
-    /* background:url(tabrightcornerunsel.gif) no-repeat scroll right top;*/
-    padding: 4px 10px 4px 10px;
-}
-
-#demo li.selected a div {
-    /*background:url(tabrightcorner.gif) no-repeat scroll right top;*/
-    padding: 4px 10px 4px 10px;
-}
-
-#demo .yui-nav li a:hover {
-    /* background: #455786 url(tableftcornerhover.gif) left top no-repeat;   */
-}
-
-#demo li.selected .nohover a:hover { /* background:#222E5D url(tableftcorner.gif)  no-repeat scroll left top; */
-}
-
-#demo li.selected .nohover a:hover div { /* background: url(tabrightcorner.gif) right top no-repeat; */
-}
-
-#demo .yui-nav li a:hover div {
-    /*background: url(tabrightcornerhover.gif) right top no-repeat; */
-}
-
-#demo .yui-content {
-    border: 1px solid #DADAD6;
-}
-
-#demo .yui-nav a {
-    border-color: #A3A3A3;
-    border-style: solid;
-    border-width: 0pt 0px;
-    color: #000000;
-    position: relative;
-    text-decoration: none;
-}
-
-#demo .yui-nav {
-    border-bottom:#1px solid #D0D0D0;
-	}
-	
-	#tab_no_link .yui-nav{
-
-	}
-	#tab_no_link .yui-nav li {
-		margin-right:0;
- 	}
-
-	#tab_no_link .yui-content { border-top-width: 2px; border-top-style: solid; border-top-color: #006699;}
-	
-	#tab_no_link .yui-nav li a { 
-			float:left;
-			cursor:text; 
-			text-decoration: none; 
-			font-size: 8pt; 
-			color:#fff;
-			background-color: white;
-			margin:0pt 0px 0pt 0pt;
-			background:#3754A1 url(tableftcornerunsel.gif) no-repeat scroll left top;
-	}
-	
-	#tab_no_link li.selected a{
-			float:left; 
-			text-decoration: none; 
-			font-size: 8pt; 
-			background:#222E5D url(tableftcorner.gif) no-repeat scroll left top;
-			color: white;
-	}
-
-   
-    #tab_no_link .longTab{
-            width:100%;
-            float:none;
-            font-size:12px;
-            color:white;
-            font-weight:bold;
-            height:25px;
-            padding-top:5px;
-            background-color:#222E5D
-			
-	}
-	#tab_no_link li a div{ 
-			background:url(tabrightcornerunsel.gif) no-repeat scroll right top;
-			padding: 4px 10px 4px 10px;
-	}
-		
-	#tab_no_link li.selected a div{ 
-			background:url(tabrightcornerunsel.gif) no-repeat scroll right top;
-			padding: 4px 10px 4px 10px;
-	}
-	#tab_no_link .yui-nav li a:hover {
-	    background: #3754A1 url(tableftcornerunsel.gif) left top no-repeat;  
-	}
-       
-        #tab_no_link li.selected .nohover a:hover {background:#222E5D url(tableftcorner.gif)  no-repeat scroll left top;}
-        
-        #tab_no_link li.selected .nohover a:hover div {background: url(tabrightcorner.gif) right top no-repeat;}
-	
-	#tab_no_link .yui-nav li a:hover div {
-	    background: url(tabrightcornerunsel.gif) right top no-repeat;
-	}
-	#tab_no_link .yui-content {
-		border-top:1px solid #000;
-	}
-	#tab_no_link .yui-nav a {
-		border-color:#A3A3A3;
-		border-style:solid;
-		border-width:0pt 0px;
-		color:#000000;
-		position:relative;
-		text-decoration:none;
-	}
-	#tab_no_link .yui-nav {
-		border-color:#222E5D;
-		border-style:solid;
-		border-width:0pt 0pt 2px;
-	}/*
-
-*/
-
