+++ amp/repository/aim/view/allVisibilityTagsComputed.jsp
@@ -886,9 +886,5 @@
 <feature:display name="Version History" module="Activity Preview"></feature:display>
 <feature:display name="Show Editable Export Formats" module="Activity Preview"></feature:display>
 
-<feature:display name="Background of organization" module="Dashboard Org. Info"></feature:display> 
-<feature:display name="Org. Description" module="Dashboard Org. Info"></feature:display> 
-<feature:display name="Key Areas of Focus" module="Dashboard Org. Info"></feature:display>
-<feature:display name="Show Names As Acronyms" module="Dashboard Org. Info"></feature:display>
 <!-- preview activity -->
 <module:display name="/Activity Form/Organizations/Donor Organization" parentModule="/Activity Form/Organizations"></module:display>
